deploy/qa-util/requirements.txt
python-gitlab==6.1.0
jira==3.8.0
requests==2.32.5
numpy==2.4.1
